Read description
Projection, lookat and translate
lookat and pos camera in GUI + camera

+++ Graphics_Engine/Graphics_Engine/Systems/Input/InputSystem.cpp
@@ -98,11 +98,11 @@ void InputSystem::Key_Callback(GLFWwindow* window, int key, int scancode, int ac
 
     if (action == GLFW_RELEASE)
     {
-      ImGuiImpl::KeyboardUp(key, _mousePosition.x, _mousePosition.y);
+      ImGuiImpl::KeyboardUp(key, (int)_mousePosition.x, (int)_mousePosition.y);
     }
     else
     {
-      ImGuiImpl::KeyboardDown(key, _mousePosition.x, _mousePosition.y);
+      ImGuiImpl::KeyboardDown(key, (int)_mousePosition.x, (int)_mousePosition.y);
     }
 }
